// tools/architecture/phase4-semantic/models/semanticKind.ts
export enum SemanticKind {

    Unknown = "Unknown",

    Class = "Class",

    Interface = "Interface",

    Function = "Function",

    Enum = "Enum",

    TypeAlias = "TypeAlias",

    Variable = "Variable",

    Component = "Component",

    Hook = "Hook",

    Repository = "Repository",

    Service = "Service",

    Controller = "Controller",

    Api = "Api",

    Page = "Page",

    Layout = "Layout",

    Storage = "Storage",

    Mapper = "Mapper",

    Engine = "Engine",

    Utility = "Utility",

    Configuration = "Configuration"

}